+++ packages/calendar-controls/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.4.0](https://github.com/schedule-x/schedule-x/compare/v4.3.1...v4.4.0) (2026-04-09)
+
+**Note:** Version bump only for package @schedule-x/calendar-controls
+
 ## [4.3.1](https://github.com/schedule-x/schedule-x/compare/v4.3.0...v4.3.1) (2026-03-09)
 
 **Note:** Version bump only for package @schedule-x/calendar-controls